requirements.txt
asgiref==3.4.1
attrs==21.4.0
certifi==2021.10.8
chardet==3.0.4
convertdate==2.3.2
export==0.1.2
fastapi==0.70.1
idna==2.8
inflection==0.5.1
jsonschema==4.0.0
html5lib==1.1
lxml==4.7.1
pandas==1.1.5
platformdirs==2.4.0
pyrsistent==0.18.0
pydantic==1.9.0
PyYAML==6.0
requests==2.22.0
routers==0.10.1
urllib3==1.25.11
uvicorn==0.16.0
virtualenv==20.13.4
opentelemetry-sdk==1.20.0
opentelemetry-api==1.20.0
opentelemetry-distro==0.41b0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
